Patch AI walk & lifebar/repair bar shown when mouse over

+++ Assets/Scripts/ScriptableObjects/Object&Map/LocationData.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public abstract class LocationData : ScriptableObject
 {
-    public float StartDist;
+    public int StartDist;
     [SerializeField]
     public GameObj_Vect2[] characterPositionList;
 }
